Conceptually, a variable has a collection of the variables to which it is connected.
This is not normalised from the API view, i.e. if a has b in its collection, b will have a in its collection.
In the XML serialisation for CellML 1.1, and proposed for CellML 1.2, this is normalised by means of the "Connection" XML element.
The implementation of the API would most likely have this normalised.