Implementation of variable connection.
Note: This explores the approach of not having a free-standing
Connection object in-memory that maps to the Connection XML element.
This has led to some interesting considerations as to how to
represent in-memory situations where the model in invalid (e.g.
while under construction, or when loaded from XML that does not
represent a valid model.)

Self-connect test still not passing.