Using MAP Client process as starting point.
As agreed at CellML meeting a few weeks ago, we will restart
the documentation of the LibCellML process using the
MAP client dev process docs as a starting point.

This is one of the docs from a family of related docs.  Some of
the others will probably also be used.